requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html

torch==1.10.0+cu111
torchvision==0.11.1+cu111
piqa
sklearn
opencv-python
matplotlib
tensorboard==2.6.0
tqdm
poppy

# From deblurGAN
ipython
PyYAML
scikit-image
pretrainedmodels
torchsummary
